fix codec bug
