+++ test/common/http/conn_manager_utility_test.cc
@@ -76,79 +76,6 @@ public:
   MOCK_METHOD(bool, isInternalAddress, (const Network::Address::Instance&), (const));
 };
 
-class MockConnectionManagerConfig : public ConnectionManagerConfig {
-public:
-  MockConnectionManagerConfig() {
-    ON_CALL(*this, generateRequestId()).WillByDefault(Return(true));
-    ON_CALL(*this, isRoutable()).WillByDefault(Return(true));
-    ON_CALL(*this, preserveExternalRequestId()).WillByDefault(Return(false));
-    ON_CALL(*this, alwaysSetRequestIdInResponse()).WillByDefault(Return(false));
-  }
-
-  // Http::ConnectionManagerConfig
-  ServerConnectionPtr createCodec(Network::Connection& connection, const Buffer::Instance& instance,
-                                  ServerConnectionCallbacks& callbacks) override {
-    return ServerConnectionPtr{createCodec_(connection, instance, callbacks)};
-  }
-
-  MOCK_METHOD(RequestIDExtensionSharedPtr, requestIDExtension, ());
-  MOCK_METHOD(const std::list<AccessLog::InstanceSharedPtr>&, accessLogs, ());
-  MOCK_METHOD(ServerConnection*, createCodec_,
-              (Network::Connection&, const Buffer::Instance&, ServerConnectionCallbacks&));
-  MOCK_METHOD(DateProvider&, dateProvider, ());
-  MOCK_METHOD(std::chrono::milliseconds, drainTimeout, (), (const));
-  MOCK_METHOD(FilterChainFactory&, filterFactory, ());
-  MOCK_METHOD(bool, generateRequestId, (), (const));
-  MOCK_METHOD(bool, preserveExternalRequestId, (), (const));
-  MOCK_METHOD(bool, alwaysSetRequestIdInResponse, (), (const));
-  MOCK_METHOD(uint32_t, maxRequestHeadersKb, (), (const));
-  MOCK_METHOD(uint32_t, maxRequestHeadersCount, (), (const));
-  MOCK_METHOD(absl::optional<std::chrono::milliseconds>, idleTimeout, (), (const));
-  MOCK_METHOD(bool, isRoutable, (), (const));
-  MOCK_METHOD(absl::optional<std::chrono::milliseconds>, maxConnectionDuration, (), (const));
-  MOCK_METHOD(absl::optional<std::chrono::milliseconds>, maxStreamDuration, (), (const));
-  MOCK_METHOD(std::chrono::milliseconds, streamIdleTimeout, (), (const));
-  MOCK_METHOD(std::chrono::milliseconds, requestTimeout, (), (const));
-  MOCK_METHOD(std::chrono::milliseconds, requestHeadersTimeout, (), (const));
-  MOCK_METHOD(std::chrono::milliseconds, delayedCloseTimeout, (), (const));
-  MOCK_METHOD(Router::RouteConfigProvider*, routeConfigProvider, ());
-  MOCK_METHOD(Config::ConfigProvider*, scopedRouteConfigProvider, ());
-  MOCK_METHOD(const std::string&, serverName, (), (const));
-  MOCK_METHOD(HttpConnectionManagerProto::ServerHeaderTransformation, serverHeaderTransformation,
-              (), (const));
-  MOCK_METHOD(ConnectionManagerStats&, stats, ());
-  MOCK_METHOD(ConnectionManagerTracingStats&, tracingStats, ());
-  MOCK_METHOD(bool, useRemoteAddress, (), (const));
-  const Http::InternalAddressConfig& internalAddressConfig() const override {
-    return *internal_address_config_;
-  }
-
-  MOCK_METHOD(bool, unixSocketInternal, ());
-  MOCK_METHOD(uint32_t, xffNumTrustedHops, (), (const));
-  MOCK_METHOD(bool, skipXffAppend, (), (const));
-  MOCK_METHOD(const std::string&, via, (), (const));
-  MOCK_METHOD(Http::ForwardClientCertType, forwardClientCert, (), (const));
-  MOCK_METHOD(const std::vector<Http::ClientCertDetailsType>&, setCurrentClientCertDetails, (),
-              (const));
-  MOCK_METHOD(const Network::Address::Instance&, localAddress, ());
-  MOCK_METHOD(const absl::optional<std::string>&, userAgent, ());
-  MOCK_METHOD(const Http::TracingConnectionManagerConfig*, tracingConfig, ());
-  MOCK_METHOD(Tracing::HttpTracerSharedPtr, tracer, ());
-  MOCK_METHOD(ConnectionManagerListenerStats&, listenerStats, ());
-  MOCK_METHOD(bool, proxy100Continue, (), (const));
-  MOCK_METHOD(bool, streamErrorOnInvalidHttpMessaging, (), (const));
-  MOCK_METHOD(const Http::Http1Settings&, http1Settings, (), (const));
-  MOCK_METHOD(bool, shouldNormalizePath, (), (const));
-  MOCK_METHOD(bool, shouldMergeSlashes, (), (const));
-  MOCK_METHOD(bool, shouldStripMatchingPort, (), (const));
-  MOCK_METHOD(envoy::config::core::v3::HttpProtocolOptions::HeadersWithUnderscoresAction,
-              headersWithUnderscoresAction, (), (const));
-  MOCK_METHOD(const LocalReply::LocalReply&, localReply, (), (const));
-
-  std::unique_ptr<Http::InternalAddressConfig> internal_address_config_ =
-      std::make_unique<DefaultInternalAddressConfig>();
-};
-
 const Http::LowerCaseString& traceStatusHeader() {
   static Http::LowerCaseString header("x-trace-status");
   return header;
