+++ test/mocks/http/BUILD
@@ -50,6 +50,7 @@ envoy_cc_mock(
         "//include/envoy/http:filter_interface",
         "//include/envoy/ssl:connection_interface",
         "//include/envoy/tracing:http_tracer_interface",
+        "//source/common/http:conn_manager_config_interface",
         "//source/common/http:filter_manager_lib",
         "//source/common/http:header_map_lib",
         "//test/mocks/event:event_mocks",
